DPD-1428: Added exception test for iframe rewritten by script
Added test which assures that blocking and allowlisting works
for the iframe content written by a script.
During the test an empty iframe is loaded and then javascript
writes a real content into this iframe and the content should
be blocked/hidden by first two filters, then allowlisted by
a third filter.

Signed-off-by: Krystian Zlomek <k.zlomek@eyeo.com>