DPD-2179: Ensure the test fails if a filter blocks a legit element
Previously, if the snippet blocked the legit element, the span that says
"Should not be hidden" disappeared, but the test remained visibly green,
there were no indicated failures.

Now, we observe changes to the element and insert a "fail" label to the
page if the element becomes hidden.

To verify, adjust the "Legit elements" filter to match:

hide-if-has-and-matches-style a[href="#comments-target-ad"] div[id] span.label ';' /\\bdisplay:\ none\;/